+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.2] - 2023-02-04
+
+## Changed
+
+- State changes are now immediate.
+- Cleaned up code to mirror changes in `LibrarIoh.StateMachines`.
+
 ## [0.1.1] - 2023-01-31
 
 ## Changed